+++ route_distances/ted/utils.py
@@ -5,10 +5,13 @@ import random
 from enum import Enum
 from operator import itemgetter
 
+import numpy as np
+from rdkit import Chem, DataStructs
+from rdkit.Chem import AllChem
 from apted import Config as BaseAptedConfig
 from scipy.spatial.distance import jaccard as jaccard_dist
 
-from route_distances.utils.type_utils import StrDict
+from route_distances.utils.type_utils import StrDict, Callable
 
 
 class TreeContent(str, Enum):
@@ -27,12 +30,19 @@ class AptedConfig(BaseAptedConfig):
 
     :param randomize: if True, the children will be shuffled
     :param sort_children: if True, the children will be sorted
+    :param dist_func: the distance function used for renaming nodes, Jaccard by default
     """
 
-    def __init__(self, randomize: bool = False, sort_children: bool = False) -> None:
+    def __init__(
+        self,
+        randomize: bool = False,
+        sort_children: bool = False,
+        dist_func: Callable[[np.ndarray, np.ndarray], float] = None,
+    ) -> None:
         super().__init__()
         self._randomize = randomize
         self._sort_children = sort_children
+        self._dist_func = dist_func or jaccard_dist
 
     def rename(self, node1: StrDict, node2: StrDict) -> float:
         if node1["type"] != node2["type"]:
@@ -40,7 +50,7 @@ class AptedConfig(BaseAptedConfig):
 
         fp1 = node1["fingerprint"]
         fp2 = node2["fingerprint"]
-        return jaccard_dist(fp1, fp2)
+        return self._dist_func(fp1, fp2)
 
     def children(self, node: StrDict) -> List[StrDict]:
         if self._sort_children:
@@ -50,3 +60,50 @@ class AptedConfig(BaseAptedConfig):
         children = list(node["children"])
         random.shuffle(children)
         return children
+
+
+class StandardFingerprintFactory:
+    """
+    Calculate Morgan fingerprint for molecules, and difference fingerprints for reactions
+
+    :param radius: the radius of the fingerprint
+    :param nbits: the fingerprint lengths
+    """
+
+    def __init__(self, radius: int = 2, nbits: int = 2048) -> None:
+        self._fp_params = (radius, nbits)
+
+    def __call__(self, tree: StrDict, parent: StrDict = None) -> None:
+        if tree["type"] == "reaction":
+            if parent is None:
+                raise ValueError(
+                    "Must specify parent when making Morgan fingerprints for reaction nodes"
+                )
+            self._add_rxn_fingerprint(tree, parent)
+        else:
+            self._add_mol_fingerprints(tree)
+
+    def _add_mol_fingerprints(self, tree: StrDict) -> None:
+        if "fingerprint" not in tree:
+            mol = Chem.MolFromSmiles(tree["smiles"])
+            rd_fp = AllChem.GetMorganFingerprintAsBitVect(mol, *self._fp_params)
+            tree["fingerprint"] = np.zeros((1,), dtype=np.int8)
+            DataStructs.ConvertToNumpyArray(rd_fp, tree["fingerprint"])
+        tree["sort_key"] = "".join(f"{digit}" for digit in tree["fingerprint"])
+        if "children" not in tree:
+            tree["children"] = []
+
+        for child in tree["children"]:
+            for grandchild in child["children"]:
+                self._add_mol_fingerprints(grandchild)
+
+    def _add_rxn_fingerprint(self, node: StrDict, parent: StrDict) -> None:
+        if "fingerprint" not in node:
+            node["fingerprint"] = parent["fingerprint"].copy()
+            for reactant in node["children"]:
+                node["fingerprint"] -= reactant["fingerprint"]
+        node["sort_key"] = "".join(f"{digit}" for digit in node["fingerprint"])
+
+        for child in node["children"]:
+            for grandchild in child.get("children", []):
+                self._add_rxn_fingerprint(grandchild, child)